Restructure API, move Vector3d and Quaternion classes
