+++ orix/io/__init__.py
@@ -16,14 +16,7 @@
 # You should have received a copy of the GNU General Public License
 # along with orix.  If not, see <http://www.gnu.org/licenses/>.
 
-"""Load and save utilities.
-
-.. warning::
-
-   These functions are far from complete or universally useful. Use at your
-   own risk.
-
-"""
+"""Load and save utilities."""
 
 import os
 from warnings import warn
@@ -34,6 +27,15 @@ import numpy as np
 from orix.io.plugins import plugin_list
 
 extensions = [plugin.file_extensions for plugin in plugin_list if plugin.writes]
+
+
+# Lists what will be imported when calling "from orix.io import *"
+__all__ = [
+    "loadang",
+    "loadctf",
+    "load",
+    "save",
+]
 
 
 def loadang(file_string):
