--- orix/objects.py
@@ -1,25 +1,0 @@
-# -*- coding: utf-8 -*-
-# Copyright 2018-2020 the orix developers
-#
-# This file is part of orix.
-#
-# orix is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# orix is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with orix.  If not, see <http://www.gnu.org/licenses/>.
-
-from orix.quaternion.orientation import Orientation, Misorientation  # pragma: no cover
-from orix.vector.neo_euler import (
-    NeoEuler,
-    Homochoric,
-    Rodrigues,
-    AxAngle,
-)  # pragma: no cover
