+++ orix/sampling/__init__.py
@@ -15,3 +15,15 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with orix.  If not, see <http://www.gnu.org/licenses/>.
+
+"""Module for generating grids in orientation spaces."""
+
+from orix.sampling.sample_generators import get_sample_fundamental, get_sample_local
+from orix.sampling.sampling_utils import uniform_SO3_sample
+
+# Lists what will be imported when calling "from orix.sampling import *"
+__all__ = [
+    "get_sample_fundamental",
+    "get_sample_local",
+    "uniform_SO3_sample",
+]
