+++ orix/scalar/__init__.py
@@ -21,10 +21,13 @@
 As well as themselves representing physical quantities, Scalars
 may represent elements of higher-dimensional quantities, such as the x-component
 of a vector or the rotation angle of a quaternion.
-
 """
+
 import numpy as np
-from orix.base import Object3d, DimensionError
+from orix.base import Object3d
+
+
+__all__ = ["Scalar"]
 
 
 class Scalar(Object3d):
